requirements.txt
geopy
plotly
matplotlib
py-moneyed
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
branca==0.3.1
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-compat==1.0.15
folium==0.10.0
idna==2.8
importlib-metadata==1.3.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
numpy==1.17.4
progressbar2==3.47.0
pyaml==19.12.0
python-utils==2.3.0
pytz==2019.3
PyYAML==5.3
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
